app/requirements.txt
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
absl-py
cycler==0.10.0
decorator==4.4.2
entrypoints==0.3
fastapi==0.63.0
gast==0.2.2
google-pasta==0.2.0
h11==0.12.0
h5py
idna==2.10
importlib-metadata 
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
Keras-Applications 
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
Markdown 
MarkupSafe==1.1.1
mistune==0.8.4
# mkl-fft==1.2.0
# mkl-random==1.1.1
# mkl-service==2.3.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.5
opt-einsum==3.1.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.13.0
ptyprocess==0.7.0
pycparser==2.20
pydantic==1.7.3
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-multipart==0.0.5
pyzmq==22.0.2
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
scipy
Send2Trash==1.5.0
six
starlette==0.13.6
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp
uvloop
httptools
websockets